refactor(discord): extract prompt-too-long terminal guidance
