+++ src/services/discord/turn_bridge/terminal_outcome_delivery.rs
@@ -33,6 +33,7 @@ use recovery_retry::{
 mod cancel_prompt_replace;
 mod delivery_epilogue;
 mod empty_response_recovery;
+mod prompt_too_long_guidance;
 mod recovery_retry;
 
 use super::*;
